Combine and remove excess components
